victara: Add a power hal ext library to manage go_hispeed_load
Victara is very responsive when interactive but tends to chew
through a lot of battery when performing background tasks (e.g.
playing music over bluetooth).

Motorola configured a very aggressive go_hispeed and this seems to
really make the interface nice.  Let's leave that alone for now
but turn it off when the screen is off.

Change-Id: I842d5c0fb3e55fb2c8cadddcc75abc4c7da71776