Update patch set 1
Patch Set 1:

I was afraid of that...  At least two moto devices support a compressed DT and it makes a huge difference (2mb compressed down to <500k and I only have a 10mb recovery partition, so that is a huge difference!).

I'm not sure how much I need this yet though.

Patch-set: 1