Update patch set 3
Patch Set 3:

If tuning number is not correct, it will not set the waking at all (check "jni_CMActions" in logcat, incorrect tuning number would be "Found tuning number X, but expected Y".

If 7 is correct, and it doesn't wake sometimes and should (need to be compared with stock), then we need to put more gestures to IR_GESTURES_FOR_SCREEN_OFF and track them as well, but right now, APPROACH and SWIPE seem to work reasonably well for me.

Patch-set: 3