Update patch set 3
Patch Set 3:

Anderson,

Can you please let me know what happens with a tuning of 6?  Can you do the following to test it?

adb shell logcat -c

unplug your phone and let it sit for a couple of minutes

test the ir sensor for wakeup and make sure to give plenty of time for the phone to enter deep sleep each time

and then post a link to the logcat on this change?  Right now the log should be reporting errors when it tries to read the ir config and after changing the tuning to 6 it should stop reporting those errors.

Patch-set: 3