Update patch set 3
Patch Set 3:

I added a new commit that allows either 6 or 7 as the tuning number (depends on this commit).  I broke it out as a separate commit to make it easy for Martin to keep this part of our code in sync if he wants.

Not that I know whether or not the commit works or breaks anything though...

Patch-set: 3