Update patch set 3
Patch Set 3:

Anderson,

I just added a little extra logging (and some logic to only enable/disable the sensor once for each call sequence and to only call silenceRinger when it is ringing).  Can you try updating to this and posting a logcat of a call that rings and is silenced and a logcat of one that doesn't ring at all?

FYI,I don't think we are missing a permission if it worked at all to silence a call for you because we have to have gotten the call state event to have enabled the ir sensor at all...

Patch-set: 3