Update patch set 4
Patch Set 4:

This actually looks like it's working to me.

It looks like the last call had a swipe very quickly (even before the screen was on).  Doing a grep CMActions shows that each call except the last one does:

Ringing, Screen is ON, event

but the last call does

Ringing, event, Screen is ON

I wonder if this is just a usability issue where it shouldn't trigger a silence in the first X ms after the ringing starts?

Patch-set: 4