Update patch set 5
Patch Set 5:

Anderson,

I added a 500ms period of time between the start of ringing and the ability for a swipe event to silence the ringer.  Can you try that and see if it works better for you?

Assuming it still doesn't ring for you, can you grab another logcat (can you add -v time to the command line to get the timestamps of the messages too?)?

Patch-set: 5