victara: cmactions: Add "Attentive Display"
Use the IR Gesture sensor to detect when an object is in front
of the screen (while it is on). As long as an object is in front of
the screen, do not let the screen timeout.

When an object is detected, use the stow sensor as a fail-safe
against leaving your phone on indefinitely in your pocket.

If the stow sensor is triggered while an object is detected and
then the sensor reports unstowed, we do not make any attempt
to resume the display lock.  This could be addressed if it
seems to be a problem in practice.

Change-Id: I8db698bc0e6434e6fe584a8739b5349728c663db