victara: cmactions: User-aware display
Use the IR Gesture sensor to detect when an object is in front
of the screen (while it is on). As long as an object is in front of
the screen, do not let the screen timeout.

When an object is detected, use the stow sensor as a fail-safe
against leaving your phone on indefinitely in your pocket.

If the phone is stowed and/or the object disappears after the
screen timeout expires, keep the display on (but dimmed) for 3 seconds
to give the user feedback that the screen is about to turn off.

Change-Id: I8db698bc0e6434e6fe584a8739b5349728c663db