Update patch set 3
Patch Set 3:

Steve,

That sounds great, but it also sounds like a big job.  We'd need to define a new (and generic) HAL for the gestures and figure out how to abstract the other sensors (stow, camera gesture, "chop-chop").

What other devices are shipping with the moto-style sensors on them?  Maybe that would help me understand how hard or easy it would be generalize this...

Patch-set: 3