Update patch set 3
Patch Set 3:

Palmer, I have an idea. If the user is using phone and scroll down page with finger and the finger cross the Stown sensor, wakelock is Released until IR sensors acquire wakelock again and for it, object need gone and back again. To solve it, you can put a delay to recognize the stown sensor.

Ex. Screen is on, keep stown sensor for Release wakelock enabled 5 seconds. This time is good to look if phone is in pocket. After 5 seconds, stown sensor can't release wakelock anymore until screen off and on again, and with this, avoid user to Release wakelock crossing the finger in stown sensor. I have it done here. I can upload the patch set or you do it if you liked and agree with me.

Patch-set: 3