Update patch set 3
Patch Set 3:

Anderson,

I was worried about the accidental stow problem that you are describing.  I thought we might need to implement a more complex state which would track the 4 possibilities of whether or not there is an object detected and whether or not it was stowed.

  Stowed?   IR Object? => WakeLock Enabled
   false       false         false
   false       true          true
   true        false         false
   true        true          false

Basically, if you have an IR object, stow the phone and then unstow it then it should go back to the same state as if it had just seen the ir object.

Any opinions on how that compares to what you have right now?

FYI, I'm still trying to understand how your unregisterListener change could fix the problem for you.  Do you understand why that fixed the problem?

Patch-set: 3