Update patch set 7
Patch Set 7:

We have something that resembles a HAL for gestures. It's really just an abstraction that translates to key events. We use it on the Oppo/OnePlus devices and the code is in device/oppo/common. 

Android added 3 new gesture types in L- WAKE_GESTURE, GLANCE_GESTURE, and PICK_UP_GESTURE. Only PICK_UP_GESTURE is actually handled (it causes ambient display to pulse). 

I implemented this using DSPS gesture detection on the Bacon. DSPS can do a bunch of other stuff, I've been meaning to detect a gesture that wakes the device and launches the camera so this would be extremely welcome in the framework. You can add vendor-specific sensor identifiers as well. These are "trigger" sensors that fire once then reset, and are all capable of waking the device from sleep.

I think there is also overlap with the proximity-pocket-detection code that we have in the framework.

Adding a few of the guys who worked on this code.

Patch-set: 7