Update patch set 7
Patch Set 7: Code-Review+1

Now it's fine. You can kill extra unregister. But I have one issue. Sometimes, when I put the phone near the wall or sometimes with nothing around, this still recognizing an object and screen still on forever. I can rec a example for you if you want. Because of this, there would use the front camera to do this instead of IR sensors, like in stock ROM?

Patch-set: 7