victara: cmactions: Fix abstraction for wake vs. sensors
I misunderstood how the JNI was configuring the sensors.  I thought
you independently confired whether or not the sensors were going to
wake the device and the sensors to be triggered.

Instead, we really configure whether or not the sensors are enabled
at all and, if they are configured, whether or not they cause the
device to wake up.

Change-Id: Ieedc39fdd69e0dd48e16cd72f5712febca636194