Update patch set 10
Patch Set 10:

Actually, giving this a second thought, I myself am rather unsure how well is this implementation of user-aware display a practical thing - activity where you look on the display has a wakelock, activity where you'd wave your hand above the display would also have a wake lock, and this would only kick in if you keep your hand above the phone and the phone would not expect interacting with it. I dunno, but it's more like a scenario where screen should go off. To me, it sounds more like a somekind of "smart unlock" gesture is practical - i.e. if your screen goes off, then for a certain perion you can only need to wave your hand to turn it back on and bypass the lock screen (if you didn't explicitly turn it off). And with swiping with hand when the screen is dimmed to get it back on.

Patch-set: 10