Update patch set 10
Patch Set 10:

You need to implement it and then try both before and after.

Set your screen timeout to the min (15 secs?) and then try to use it both with and without the option enabled.

For me, with the option disabled I frequently have the display turn off while I am waiting for slow remote sites or services.

With the option enabled the net result is I can use things like I normally do with a very low timeout.

The big negative of this is that it can keep your screen on for too long.  For example, where I normally charge my phone at night there is a plant and it registers as an object...

Patch-set: 10