Update patch set 1
Patch Set 1: Code-Review-2

I added this on mondrianwifi (Slim was going to support it and I didn't want to make it hard on people to switch) and everyone immediate made both partitions f2fs on the assumption that it is better even though I (and motorola for that matter) believe it is not.

I prefer to not give people the option of making things worse for themselves if they are going to blindly do so with no real clue.

Patch-set: 1
Label: Code-Review=-2