Update patch set 1
Patch Set 1: Code-Review-2

I think I went down this rat hole before and a real problem is that you cannot take a device that was software encrypted and then install a build with hardware encryption on it.

We would need to verify that it worked now or else work on patches to system/core to do so.

Patch-set: 1
Label: Code-Review=-2