Update patch set 1
Patch Set 1:

No, I have not.

At the very least the wcnss nv bin needs to be symlinked into /data/whatever. Most devices just ship this blob on /system/ and then do the symlink to /data/. We're symlinking it from /persist/ which I think we should avoid because we'll have more control over it being in our repo.

Patch-set: 1