Update patch set 1
Patch Set 1:

Micah, thanks for the clarification.  That was the big disconnect (and kind of a funny one).

I didn't realize that motorola has started to partially use the XML to configure this.  It's only partial though because the speaker changes are based on source code changes they made (I reversed the table in audio.primary.msm8974.so and made the changes in our XML).

Thanks to both of you for pointing it out.  Can you please test the new commit I just uploaded (I can't test it) which combines the two.

Patch-set: 1