Update patch set 2
Patch Set 2:

I was looking into the libjustshoot.so, and it seems that the _ZN7android8AMessageC1Eji is called a lot in this lib.
Are all the methods from libjustshoot.so not used, so we can only define an empty body for those methods?

Patch-set: 2
Reviewer: Benjamin Legrand <12857@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0