Update patch set 1
Patch Set 1:

The first (only?) problem using the source built ril was that

responseDataCallList

uses its own RIL_Data_Call_Response_v9caf.  This structure looks roughly the same as the v11 that we are using except that they do not use the "status" field at all and there is 1 extra byte at the start of the structure (wtf?).

I am inclined to want to start with the proprietary version at this time.  Updating our structure to match the bytes is risky because we have no evidence that the bytes which would have to call 'status' actually contain anything meanginful, let alone contain a status.  Therefore I think we'd need to import the whole libril source and modify it to safely support our radio which we can always do in the future if this is a problem.

Patch-set: 1