Update patch set 3
Patch Set 3:

"Inside" mpdecision, the original default boost time was 500ms. My 500ms were a workaround for our janky framework in L and interactive wasn't as great as it is now. And we have the app start boost on top now, there are no fluctuations and stutters anymore.

After seeing Dan's commit yesterday, I played with qcom's 40ms boost and I'm enclined to think, this is sufficient. I gues I'm reverting as well. So thanks for this heads-up.

Patch-set: 3