Update patch set 1
Patch Set 1:

In my car this morning it didn't auto-connect to the head unit.  Manually connecting (not pairing, it's already paired) took 3 or 4 tries before it finally connected.

After this change I just tried it twice and worked both times to auto-connect when I turned the car on.

Pairing has other wonkiness too (I sometimes have to refresh 3 or 4 times to see the device that I'm trying to pair with and then try pairing a few times).

I haven't really tested that to see if it is fixed by it.

Not that I'm entirely sure how this would fix bluetooth though.

Patch-set: 1