victara: Finish overloading recoveryimg
assert-max-image-size has recently been failing in my experience due
to the usage of the standard build-recoveryimage-target routine. This
routine calls assert-max-image-size on the output of
  $(MKBOOTIMG) $(INTERNAL_RECOVERYIMAGE_ARGS) $(BOARD_MKBOOTIMG_ARGS)
which does not pickup the LZMA compressed ramdisk. Since MKBOOTIMG is
already handled in victara's overloaded recovery image generation,
just check its size after creation and be done with it.

Remove unsupported 3rd argument to assert-max-image-size in the boot
image generation as well.

Change-Id: I4ed8661c6af6317361992e65a2f88188726b8eb7