Update patch set 1
Patch Set 1: Code-Review-1

Placebo effect?  Or does it magically work somewhat because the kernel's error checking for invalid sample rates is bad?

Looking at the code, I think to enable that setting you would also need to have this commit in the kernel:

http://review.cyanogenmod.org/141937

fwiw, I compiled the commit but haven't booted with it (it needs your commit for the hal's boardflag or will completely bork audio).

Patch-set: 1