Update patch set 1
Patch Set 1:

This looks like crap because we force hdpi graphics for recovery to save space in recovery which then forces the cm recovery images to also be small and crappy.

We could change vendor/cm to let us override that, or add our own images to our device tree but I'm inclined to leave it as-is and call it functional.

Patch-set: 1