Update patch set 1
Patch Set 1:

> How does this "decrease the occurrences of battery shutdown at
 > ~50%" ?

This is just a guess (I worded it wrong). But according to the Android.com page, the purpose of this file is to "estimate battery drain correctly", and the previous cpu.active values were a bit high when compared against the stock file. 

I thought that maybe having a high current value for the different CPU frequencies could be related to these early shutdowns, but I don't think this alone will fix the issue.

Patch-set: 1
Label: Verified=0