Update patch set 1
Patch Set 1:

I'm not sure if I should use CM or just regular charger images in the .Mk files. What's the difference, and, the way it is setup at the moment, would it replace the current ugly charger image that we have, when the phone is charging but not turned on?

Patch-set: 1
Reviewer: Alexandre Oliveira <15485@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>