Update patch set 3
Patch Set 3:

this helped for 8084
https://review.lineageos.org/#/c/LineageOS/android_device_samsung_apq8084-common/+/220301/

Which is better than this.

And yes there is not solution only workarounds, I don't see the problem happening anymore on my device or complains about it be a issue in threads for a long time, the last attempt I did was using that change, but as the device are different others specif changes may also helped 8084.

Patch-set: 3
CC: Felipe Leon <14611@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>