Update patch set 1
Patch Set 1:

Yeah, sentiment I got was this is basically this is a no-go.

I realize it was merged on hh 16.0, I'll revert that at some point, as it shouldn't have been merged, it was accidentally merged in a push.

Patch-set: 1
Reviewer: Gerrit User 16185 <16185@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>